Add timer utility and update HomePage and Footer components
This commit introduces a timer utility which formats and displays the seconds passed since a specified start date. The HomePage component has been updated with new sections and the use of the timer utility. The Footer component has also been updated with a new stylized look and the addition of the current year.